// src/generated/models/DeploymentUnitCreationRequest.ts
export interface DeploymentUnitCreationRequest {
    /**
     * 
     * @type {number}
     * @memberof DeploymentUnitCreationRequest
     */
    stage: number;
    /**
     * 
     * @type {number}
     * @memberof DeploymentUnitCreationRequest
     */
    cluster: number;
    /**
     * 
     * @type {number}
     * @memberof DeploymentUnitCreationRequest
     */
    lane: number;
    /**
     * 
     * @type {string}
     * @memberof DeploymentUnitCreationRequest
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof DeploymentUnitCreationRequest
     */
    namespace: string;
    /**
     * 
     * @type {string}
     * @memberof DeploymentUnitCreationRequest
     */
    template?: string | null;
    /**
     * 
     * @type {number}
     * @memberof DeploymentUnitCreationRequest
     */
    replicas?: number;
    /**
     * 
     * @type {number}
     * @memberof DeploymentUnitCreationRequest
     */
    cpuLimit?: number;
    /**
     * 
     * @type {number}
     * @memberof DeploymentUnitCreationRequest
     */
    memoryLimit?: number;
    /**
     * 
     * @type {number}
     * @memberof DeploymentUnitCreationRequest
     */
    cpuRequest?: number;
    /**
     * 
     * @type {number}
     * @memberof DeploymentUnitCreationRequest
     */
    memoryRequest?: number;
    /**
     * 
     * @type {number}
     * @memberof DeploymentUnitCreationRequest
     */
    progressDeadline?: number;
}