// src/generated/models/DeploymentUnitMutationRequest.ts
export interface DeploymentUnitMutationRequest {
    /**
     * 
     * @type {number}
     * @memberof DeploymentUnitMutationRequest
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof DeploymentUnitMutationRequest
     */
    stage?: number;
    /**
     * 
     * @type {number}
     * @memberof DeploymentUnitMutationRequest
     */
    cluster?: number;
    /**
     * 
     * @type {number}
     * @memberof DeploymentUnitMutationRequest
     */
    lane?: number;
    /**
     * 
     * @type {string}
     * @memberof DeploymentUnitMutationRequest
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof DeploymentUnitMutationRequest
     */
    namespace?: string;
    /**
     * 
     * @type {string}
     * @memberof DeploymentUnitMutationRequest
     */
    template?: string | null;
    /**
     * 
     * @type {number}
     * @memberof DeploymentUnitMutationRequest
     */
    replicas?: number;
    /**
     * 
     * @type {number}
     * @memberof DeploymentUnitMutationRequest
     */
    cpuLimit?: number;
    /**
     * 
     * @type {number}
     * @memberof DeploymentUnitMutationRequest
     */
    memoryLimit?: number;
    /**
     * 
     * @type {number}
     * @memberof DeploymentUnitMutationRequest
     */
    cpuRequest?: number;
    /**
     * 
     * @type {number}
     * @memberof DeploymentUnitMutationRequest
     */
    memoryRequest?: number;
    /**
     * 
     * @type {number}
     * @memberof DeploymentUnitMutationRequest
     */
    progressDeadline?: number;
}